Pull out a big chunk of profiler logic into its own file
